Refactor PJE bot structure: separate resources into dedicated modules
- Moved `capa` and `resources` imports to a dedicated `resources` module.
- Created new `resources` module with `autenticador`, `buscador`, `formatador`, and `varas_dict`.
- Updated import paths in `movimentacao.py` and `pauta.py` to reflect new resource structure.
- Implemented `capa.py` with core functionality for the PJE bot.
- Enhanced `autenticador.py` to handle authentication with detailed error handling.
- Added `buscador.py` for process searching and captcha resolution.
- Introduced `formatador.py` for data formatting tasks.
- Established `varas_dict.py` for mapping court branch names to codes.
