+++ crawjud/bots/pje/pauta.py
@@ -18,7 +18,7 @@ from selenium.webdriver.common.by import By
 from selenium.webdriver.remote.webelement import WebElement
 from selenium.webdriver.support import expected_conditions as ec
 
-from crawjud.bots.pje.res.varas_dict import varas as varas_pje
+from crawjud.bots.pje.resources.varas_dict import varas as varas_pje
 from crawjud.common.exceptions.bot import ExecutionError
 from crawjud.core import CrawJUD
