+++ crawjud/bots/pje/resources/varas_dict.py
